+++ CHANGELOG.md
@@ -18,6 +18,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   were not converted to a serializable form. Now, `Quantity` are converted to `str` in
   `as_dict()`.
 
+### Removed
+
+- python 3.8 is no longer supported
+
 ## [0.8.1] - 2023-10-01
 
 ### Changed