+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Solution.components` is now automatically sorted in descending order of amount, for
   consistency with `anions`, `cations`, and `neutrals`.
 
+### Removed
+
+- `Solution.list_solutes()` has been removed. Use `Solution.components.keys()` instead.
+
 ### Fixed
 
 - Bugfix in `as_dict` to make serialization via `dumpfn` possible. Previously, `Quantity`